feat: Migrate tracer analysis tests
